Update generator logic and location for rtkq folder
